software/requirements.txt
dynaconf==3.2.11
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pygame==2.6.1
Pygments==2.19.2
pyserial==3.5
pytest==8.4.1
